playback: return message if recording is not enable or there's no running recording
