+++ packages/api/src/controllers/playback.ts
@@ -19,6 +19,7 @@ import { WithID } from "../store/types";
 import { NotFoundError, UnprocessableEntityError } from "../store/errors";
 import { isExperimentSubject } from "../store/experiment-table";
 import logger from "../logger";
+import { getRunningRecording } from "./session";
 
 /**
  * CROSS_USER_ASSETS_CUTOFF_DATE represents the cut-off date for cross-account
@@ -38,7 +39,9 @@ function newPlaybackInfo(
   playbackPolicy?: Asset["playbackPolicy"] | Stream["playbackPolicy"],
   staticFilesPlaybackInfo?: StaticPlaybackInfo[],
   live?: PlaybackInfo["meta"]["live"],
-  recordingUrl?: string
+  recordingUrl?: string,
+  withRecordings?: boolean,
+  streamRecord?: boolean
 ): PlaybackInfo {
   let playbackInfo: PlaybackInfo = {
     type,
@@ -73,14 +76,25 @@ function newPlaybackInfo(
       url: webRtcUrl,
     });
   }
-  if (recordingUrl) {
+  if (withRecordings) {
     playbackInfo.meta.dvrPlayback = [];
-    playbackInfo.meta.dvrPlayback.push({
-      hrn: "HLS (TS)",
-      type: "html5/application/vnd.apple.mpegurl",
-      url: recordingUrl,
-    });
+    if (!streamRecord) {
+      playbackInfo.meta.dvrPlayback.push({
+        message: "Error: Recording is not enabled for this stream.",
+      });
+    } else if (recordingUrl) {
+      playbackInfo.meta.dvrPlayback.push({
+        hrn: "HLS (TS)",
+        type: "html5/application/vnd.apple.mpegurl",
+        url: recordingUrl,
+      });
+    } else {
+      playbackInfo.meta.dvrPlayback.push({
+        message: "Error: No running recordings available.",
+      });
+    }
   }
+
   return playbackInfo;
 }
 
@@ -195,7 +209,7 @@ async function getPlaybackInfo(
   isCrossUserQuery: boolean,
   origin: string,
   withRecordings?: boolean,
-  recordCatalystObjectStoreId?: string
+  req?: Request
 ): Promise<PlaybackInfo> {
   const cutoffDate = isCrossUserQuery ? null : CROSS_USER_ASSETS_CUTOFF_DATE;
   let { stream, asset, session } = await getResourceByPlaybackId(
@@ -218,12 +232,10 @@ async function getPlaybackInfo(
   }
 
   if (stream) {
-    let recordingPlaybackUrl: string;
-    if (withRecordings) {
-      recordingPlaybackUrl = await getRecordingPlaybackUrl(
-        stream,
-        recordCatalystObjectStoreId
-      );
+    let url: string;
+    let fetchRunningRecording = withRecordings && stream.record;
+    if (fetchRunningRecording) {
+      ({ url } = await getRunningRecording(stream, req));
     }
     return newPlaybackInfo(
       "live",
@@ -232,7 +244,9 @@ async function getPlaybackInfo(
       stream.playbackPolicy,
       null,
       stream.isActive ? 1 : 0,
-      recordingPlaybackUrl
+      url,
+      fetchRunningRecording,
+      stream.record
     );
   }
 
@@ -274,7 +288,7 @@ app.get("/:id", async (req, res) => {
     isEmbeddablePlayer,
     origin,
     withRecordings,
-    req.config.recordCatalystObjectStoreId
+    req
   );
   if (!info) {
     throw new NotFoundError(`No playback URL found for ${id}`);
